--- src/test/resources/HW18.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-
-<suite name="TestNG">
-
-    <test name="HW-18" preserve-order="false">
-        <classes>
-            <!--<class name="LoginTests"/>-->
-            <class name="BaseTest"/>
-            <class name="Homework18"/>
-        </classes>
-    </test>
-</suite>
