+++ src/test/resources/HW20.xml
@@ -11,7 +11,7 @@
         <classes>
             <!--<class name="LoginTests"/>-->
             <class name="BaseTest"/>
-            <class name="Homework19"/>
+            <class name="Homework20"/>
         </classes>
     </test>
 </suite>
