--- src/test/resources/TestNG.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-
-<suite name="TestNG">
-
-    <test name="Example Test" preserve-order="false">
-        <classes>
-            <!--<class name="LoginTests"/>-->
-            <class name="Homework17"/>
-            <class name="BaseTest"/>
-        </classes>
-    </test>
-</suite>
